+++ packages/core-cairo/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Breaking changes**:
   - Remove non-standard `safeAllowance` option from ERC20. ([#324](https://github.com/OpenZeppelin/contracts-wizard/pull/324))
+  - Use `abi(per_item)` attribute instead of `external` for generated trait impls.
 
 ## 0.8.0 (2023-12-11)